# Add property change guards to Sprite2D, Sprite3D and AnimatedSprite2D