Implements this proposal: https://github.com/godotengine/godot-proposals/issues/8510

If a property's setter is called with the property's current value, return immediately instead of doing anything costly.

My test creates 10000 Sprite2Ds.

```gdscript
func _ready():
	for y in 100:
		for x in 100:
			var s = SPRITE.instantiate()
			s.position = Vector2(x* 10, y* 10)
			add_child(s)
```

Each sprite has a script which does some combination of these simple setters (or just `pass`).

```gdscript
func _process(delta):
	pass
	flip_h = true
	frame = 0
```

My test version was built with Visual Studio and it's performance seems to be a bit lower than the official version. "% compared to pass" numbers are calculated from release version fps numbers.

|                                 | master, debug fps | master, release fps | % compared to pass | PR, debug fps | PR, release fps | % compared to pass |
|---------------------------------|------------------:|--------------------:|-------------------:|--------------:|----------------:|-------------------:|
| `pass`                          |                73 |                 105 |               100% |            74 |              98 |               100% |
| `flip_h = true`                 |                35 |                  48 |                46% |            32 |              92 |                94% |
| `frame = 0`                     |                50 |                  95 |                91% |            60 |              91 |                93% |
| `flip_h = true` and `frame = 0` |                31 |                  47 |                45% |            30 |              89 |                91% |

It may seem odd to just to set `flip_h = true`, but it is a very common practice to set sprite's direction in every frame like this:
```gdscript
	flip_h = velocity.x > 0
```
So there are a lot of 2D games which are setting `flip_h` in every frame without actually changing it.

_Production edit: Closes https://github.com/godotengine/godot-proposals/issues/8510_