@arkology 2023-11-24
I have one really stupid style question. Suggested code follows pattern `if (class_variable == func_arg)`, so it checks if current sprite property is same as newly set. How about to check opposite, if newly set is same as sprite property - so we focus on function argument, not sprite property? Of course this is absolutely the same in terms of functionality but I think it has slight logical difference.
Is there any style recommendations about it in Godot codebase or it doesn't actually make any sense to care about? :)

(And of course I support this guards addition, without it some users could have redraw spam without even knowing about it.)