@akien-mga 2023-11-24
Yeah the order used here is the most common in my experience, and the most readable IMO.