@kleonc 2023-12-12
Could be also added to `SpriteBase3D::set_centered` etc. (in `sprite_3d.cpp`).