@AThousandShips 2023-12-12
Also unify code in `Sprite3D::set_region_rect`