@miv391 2023-12-12
> Could be also added to `SpriteBase3D::set_centered` etc. (in `sprite_3d.cpp`).

Done.

> Also unify code in `Sprite3D::set_region_rect`, and the different syntax in `SpriteBase3D::set_alpha_hash_scale` etc.

Done.

I _didn't_ add a change guard to `set_modulate`: https://github.com/godotengine/godot/blob/84692c6252438393999c427bf7f5d35b9c43926c/scene/3d/sprite_3d.cpp#L324-L328 

That function propagates the color to child nodes and I have no idea what would happen if you add new child nodes between setting modulation color like this:
```gdscript
sprite3d.modulate = Color.RED
sprite3d.add_child(some_node)
sprite3d.modulate = Color.RED
```
At least now without the change guard the color will be propagated to the child node.