@YuriSizov 2023-12-15
> That function propagates the color to child nodes and I have no idea what would happen if you add new child nodes between setting modulation color like this:

It doesn't actually propagate any values, it just triggers a redraw (which makes `queue_redraw` in this method redundant btw). It's safe to add a guard there, newly added children would of course be drawn on their own.