@YuriSizov 2023-12-16
Good work, needs a rebase after https://github.com/godotengine/godot/pull/85317. I merged it first since it fixes a problem and is generally more involved.